refactor: _compare moves to its own module
_rows will need the tolerance rule and examples will import _rows, so
leaving _compare in examples would make that a cycle. Pure move; the
suite is unchanged.
