+++ src/agentic_data_contracts/validation/examples.py
@@ -21,6 +21,7 @@ from agentic_data_contracts.adapters._normalizer import SqlNormalizer
 from agentic_data_contracts.core.contract import DataContract
 from agentic_data_contracts.validation._scalar import _scalar
 from agentic_data_contracts.validation._timewindow import _relative_time_node
+from agentic_data_contracts.validation._tolerance import _compare
 from agentic_data_contracts.validation.explain import ExplainAdapter
 from agentic_data_contracts.validation.validator import ValidationResult, Validator
 
@@ -198,37 +199,6 @@ def _fmt(value: float | None) -> str:
     crashing the whole report.
     """
     return "?" if value is None else f"{value:.3g}"
-
-
-def _compare(
-    actual: float, expected: float, rel_tol: float, abs_tol: float
-) -> tuple[float, float, bool]:
-    """Compare a measured value against a certified answer.
-
-    Returns ``(abs_diff, rel_diff, matched)``.
-
-    ``rel_diff`` is guarded against a zero reference — a certified answer of
-    zero is legitimate, and dividing by it would raise or report a meaningless
-    ``inf`` for an exact match. The three-branch form is the same one
-    ``reconcile_decomposition`` uses for a zero parent.
-
-    The relative term is anchored on ``expected``, deliberately unlike
-    ``reconcile_decomposition`` (which compares two measurements and has no
-    privileged side) and unlike ``math.isclose`` (which anchors on the larger
-    magnitude). An assertion *has* a reference: the certified answer is the
-    fixed point and the query result is what varies against it. Anchoring on
-    ``expected`` keeps the tolerance's meaning stable — "within 0.1% of the
-    certified number" — however far the query has drifted.
-    """
-    abs_diff = abs(actual - expected)
-    if abs_diff == 0:
-        rel_diff = 0.0
-    elif expected != 0:
-        rel_diff = abs_diff / abs(expected)
-    else:
-        rel_diff = math.inf
-    matched = abs_diff <= max(abs_tol, rel_tol * abs(expected))
-    return abs_diff, rel_diff, matched
 
 
 @dataclass
